# to make repetition penalty faster