--- a/vllm/model_executor/sampling_metadata.py
+++ b/vllm/model_executor/sampling_metadata.py
@@ -9,8 +9,7 @@ from vllm.sampling_params import SamplingParams, SamplingType
 from vllm.sequence import (VLLM_TOKEN_ID_ARRAY_TYPE, SequenceData,
                            SequenceGroupMetadata)
 from vllm.utils import (PyObjectCache, async_tensor_h2d,
-                        is_pin_memory_available, 
-                        make_tensor_with_pad,
+                        is_pin_memory_available, make_tensor_with_pad,
                         make_tensor_with_pad_align)
 
 _SAMPLING_EPS = 1e-5