@kdamaszk 2024-11-22
@ccrhx4 please check CI status -- there are still some formatting issues with yapf which should be easy to fix.
However, there is also more important issue with CPU tests -- please note that they are using fake_hpu (HPU path simulated on CPU). You need to cover it properly -- you can use `from vllm.utils import is_fake_hpu`.
To reproduce on local machine please run:
```
VLLM_SKIP_WARMUP=true VLLM_PROMPT_SEQ_BUCKET_MAX=128 VLLM_USE_FAKE_HPU=1 python examples/offline_inference_fakehpu.py
```