@ccrhx4 2024-12-03
@kdamaszk I see. I will update my code.