@ccrhx4 2024-12-03
@kdamaszk I will submit a new PR to include only changes in sampling logic and do not touch pin_memory. pin_memory support on HPU is not mature enough, it will make the code dirty if I follow HPU's way (only HPU goes this way.)

I am submiting JIRAs asking for pin_memory support in default pytorch way on HPU.